refactor checkbox with dependant-group trait
