+++ src/UI/Component/Input/Field/Checkbox.php
@@ -11,23 +11,6 @@ use ILIAS\UI\Component\Onloadable;
 /**
  * This describes checkbox inputs.
  */
-interface Checkbox extends Group {
+interface Checkbox extends Input {
 
-	/**
-	 * Creates a Checkbox like this but with a dependant group attached which appears if the
-	 * control is clicked.
-	 *
-	 * @param DependantGroup $dependant_group group to be attached to the checkbox
-	 *
-	 * @return Checkbox
-	 */
-	public function withDependantGroup(DependantGroup $dependant_group);
-
-
-	/**
-	 * Returns the attached DependantGroup or null if none is attached.
-	 *
-	 * @return $dependantGroup|null
-	 */
-	public function getDependantGroup();
 }
